Rename translation to translation_central
