+++ server/src/sync/translation_central/legacy_row_types.md
