+++ server/src/sync/translation_central/name.rs
@@ -1,4 +1,4 @@
-use crate::sync::translation::{SyncTranslationError, TRANSLATION_RECORD_NAME};
+use crate::sync::translation_central::{SyncTranslationError, TRANSLATION_RECORD_NAME};
 use repository::schema::{CentralSyncBufferRow, NameRow};
 
 use serde::Deserialize;
@@ -43,7 +43,7 @@ impl LegacyNameRow {
 
 #[cfg(test)]
 mod tests {
-    use crate::sync::translation::{
+    use crate::sync::translation_central::{
         name::LegacyNameRow,
         test_data::{name::get_test_name_records, TestSyncDataRecord},
     };
