+++ server/src/sync/translation_central/test_data/master_list.rs
@@ -1,4 +1,4 @@
-use crate::sync::translation::test_data::{TestSyncDataRecord, TestSyncRecord};
+use crate::sync::translation_central::test_data::{TestSyncDataRecord, TestSyncRecord};
 use repository::schema::{CentralSyncBufferRow, MasterListRow};
 
 const MASTER_LIST_1: (&'static str, &'static str) = (
