+++ server/src/sync/translation_central/test_data/master_list_line.rs
@@ -1,4 +1,4 @@
-use crate::sync::translation::test_data::{TestSyncDataRecord, TestSyncRecord};
+use crate::sync::translation_central::test_data::{TestSyncDataRecord, TestSyncRecord};
 use repository::schema::{CentralSyncBufferRow, MasterListLineRow};
 
 const MASTER_LIST_LINE_1: &'static str = r#"{
