+++ server/src/sync/translation_central/test_data/master_list_name_join.rs
@@ -1,4 +1,4 @@
-use crate::sync::translation::test_data::{TestSyncDataRecord, TestSyncRecord};
+use crate::sync::translation_central::test_data::{TestSyncDataRecord, TestSyncRecord};
 use repository::schema::{CentralSyncBufferRow, MasterListNameJoinRow};
 
 const LIST_MASTER_NAME_JOIN_1: (&'static str, &'static str) = (
