+++ server/src/sync/translation_central/unit.rs
@@ -1,4 +1,4 @@
-use crate::sync::translation::{SyncTranslationError, TRANSLATION_RECORD_UNIT};
+use crate::sync::translation_central::{SyncTranslationError, TRANSLATION_RECORD_UNIT};
 use repository::schema::{CentralSyncBufferRow, UnitRow};
 
 use serde::Deserialize;
@@ -47,7 +47,7 @@ impl LegacyUnitRow {
 
 #[cfg(test)]
 mod tests {
-    use crate::sync::translation::{
+    use crate::sync::translation_central::{
         test_data::{unit::get_test_unit_records, TestSyncDataRecord},
         unit::LegacyUnitRow,
     };
